Save peer CSRK after ltk generation
Security level at stack side is determined after the LTK generation,
saving peer csrk info prior to this would result in saving security level as ZERO for peer CSRK info in to config file.
This can cause clearing the MITM_AUTH flag when CSRK info is loaded in the next BT session as security level is saved as ZERO for CSRK. Saving the peer CSRK after ltk generation would help in saving the right security level

This is a backport of aosp/3254899 against b#362322766.  Patches need to be included in the fix for b#288144143 to prevent regressions as described in b#428948282.

Test: mmm packages/modules/Bluetooth
Tag: #security
Bug: 288144143
Merged-In: I7b9061db13cdcdbe3535c913dfd29f5a30449803
Change-Id: I61bdd1f85e05388d0058b03fdc269f9eaca86c3f